Files related to PDF loader moved to a new folder
